+++ src/PageState.ts
@@ -1,21 +1,22 @@
-import { callbacks } from "./callback";
-import { emitter } from "./InternalEmitter";
 import type { AppManager } from "./AppManager";
 import type { PageState } from "./Page";
 
+import { internalEmitter } from "./InternalEmitter";
+import { callbacks } from "./callback";
+
 export class PageStateImpl {
     constructor(private manager: AppManager) {
-        emitter.on("changePageState", () => {
+        internalEmitter.on("changePageState", () => {
             callbacks.emit("pageStateChange", this.toObject());
         });
     }
 
     public get index(): number {
-        return this.manager?.store.getMainViewSceneIndex() || 0;
+        return this.manager.store.getMainViewSceneIndex() || 0;
     }
 
     public get length(): number {
-        return this.manager?.mainViewScenesLength || 0;
+        return this.manager.mainViewScenesLength || 0;
     }
 
     public toObject(): PageState {